Move public API into its own file
