+++ src/headless/shared/api/events.js
@@ -1,4 +1,4 @@
-import _converse from '@converse/headless/shared/_converse.js';
+import _converse from '../_converse.js';
 import isFunction from '../../utils/core.js';
